requirements.txt
Flask
torch
torchvision
numpy
Pillow
gunicorn
